Indent product_ids handling code the way it seems to be intended
The code in the Vulnerability._handle_remediations method has code that
determines product IDs and binds it to the local variable product_ids.
The code that would then put its value into the remediation dict was not
indented enough so that it was executed even if the product_ids variable
hadn't been set, leading to error messages and pylint messages.
Indenting the code to the same level as the assignment to product_ids
fixes these issues.

With this change, the generated CSAF files differ in some cases and for
some inputs the converter will now find new errors. Examples:

 - examples/1.2/cvrf_example_c.xml now produces a CSAF file with a
   vulnerabilities array.

 - examples/1.2/cvrf_example_e.xml now fails with:

   ERROR - Mandatory test is_valid_defined_product_ids failed.