Improve logic regarding remediations missing product_ids/group_ids
In the CSAF output, every remediation item has to have at least one of
product_ids and group_ids (see 3.2.3.12 Vulnerabilities Property -
Remediations). If neither is present in the input CVRF, the code tries
to get the product_ids from the product_status. If that fails, the
converter should produce an error (see 9.1.5 Conformance Clause 5: CVRF
CSAF converter).

Indenting part of the logic for this in the previous commit didn't quite
get this logic right. This defect appears to have been present already
in the original wrongly indented code.